# [FVM] Deduct fees and limit storage in synthetic benchmarks